Fixed few more issues in CM path
Addressed isssues mentioned below
1. Corrected EventDataOp during removal of fru
2. Passing same new container id while updating BMC,host tree

During removal of fru, we do remove the contained entity
from entity association pdr. If this is the only contained
entity in the PDR, the entity association PDR is removed.
Based on if the entity association PDR is modified or deleted
the EventDataOp is set.

Added the check for presence of record inside PDR repo to know
if the PDR is modified and deleted.

After CM, card PDR was getting created with wrong containerId and
the connecter PDR were getting created with containerid 0. Fixed
the issue by passing same new container id while updating BMC and
host tree

Tested:
busctl commands

Signed-off-by: Archana Kakani <archana.kakani@ibm.com>